revert master organization to what it was before the last demo
