--- run_keepalive.py
@@ -1,3 +1,0 @@
-from bitmessage.bitmessage_keepalive import find_old_collections
-
-find_old_collections(3)
