--- run_listener.py
@@ -1,3 +1,0 @@
-from bitmessage.bitmessage_listener import get_collections
-
-get_collections()
