+++ tasks/add_task.py
@@ -1,4 +1,4 @@
-from webapp import celery
+from frontend.webapp import celery
 
 @celery.task()
 def add_together(a, b):
